Fix compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@23987 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/Settings/SetSavedDataPage.cpp
@@ -136,7 +136,7 @@ BOOL CSetSavedDataPage::OnInitDialog()
     CRegistryKey regCerts(L"Software\\TortoiseSVN\\CAPIAuthz");
     CStringList certList;
     regCerts.getValues(certList);
-    int nCapi = certList.GetCount();
+    int nCapi = (int)certList.GetCount();
 
     CDirFileEnum logenum(CPathUtils::GetAppDataDirectory()+_T("logcache"));
     while (logenum.NextFile(sFile, &bIsDir))
